refactor(parser): extract type resolver into types/resolve
Plan 2, step 6 of N (final extraction). Move the resolution
machinery — `resolve_arguments` / `resolve_one` / `resolve` /
`resolve_inner` / `resolve_name` / `resolve_toolr_types_name` /
`resolve_subscript` / `resolve_bin_op` / `is_count_type` /
`follow_alias_for_{path_constraints,arg_metadata}` / the small
`literal_string_values` / `tuple_element_exprs` / `render_attribute`
/ `annotation_to_label` helpers — into
`crates/toolr-core/src/parser/types/resolve.rs` (412 lines).

`resolve_toolr_types_name` promoted to `pub(super)` so the existing
`toolr_types_module_names_round_trip` test in mod.rs can still
exercise the enumerated set directly.

`types/mod.rs` shrinks 788 → 403 lines, of which ~370 are the test
module that exercises the public surface end-to-end. The rest is
just the `mod` declarations + `pub use` re-exports + the
`pub(super) is_toolr_arg_call` helper shared by `path_constraints.rs`
and `arg_metadata.rs`.

## Plan 2 final shape

| File                | Lines | Responsibility                                    |
| ------------------- | ----: | ------------------------------------------------- |
| `mod.rs`            |   403 | declarations + shared helper + test module        |
| `resolve.rs`        |   412 | annotation → SupportedType resolver pipeline      |
| `imports.rs`        |   256 | TypeImports + SourcesImports import-alias tables  |
| `arg_metadata.rs`   |   148 | extract_arg_metadata + help-section helpers       |
| `supported.rs`      |   125 | SupportedType + UnsupportedType + diag types      |
| `path_constraints.rs`|   87 | PathConstraints + extract_path_constraints        |
| `literals.rs`       |    32 | literal_str / literal_str_list / literal_u32      |

Was: one 1376-line file with 38 top-level items. Now: seven files
with focused, single-responsibility boundaries. No file is larger
than 412 lines; the avg is ~209.

Public API surface unchanged — every type and function callers
used (`parser::types::SupportedType`, `::TypeImports`, ::resolve,
::resolve_arguments, ::extract_arg_metadata, ::extract_path_constraints,
::PathConstraints, ::TypeResolutionError, ::UnsupportedType,
::SourcesImports) continues to resolve through the `pub use`
re-exports in `mod.rs`.

Tests: 24 / 24 parser::types::tests still pass; `cargo check
--workspace` clean.
